+++ examples/common/async_twisted_client.py
@@ -11,15 +11,15 @@ client implementation from pymodbus.
 # --------------------------------------------------------------------------- #
 from twisted.internet import reactor
 
-from pymodbus.client.async.tcp import AsyncModbusTCPClient
-# from pymodbus.client.async.udp import AsyncModbusUDPClient
-from pymodbus.client.async import schedulers
+from pymodbus.client.asynchronous.tcp import AsyncModbusTCPClient
+# from pymodbus.client.asynchronous.udp import AsyncModbusUDPClient
+from pymodbus.client.asynchronous import schedulers
 
 # --------------------------------------------------------------------------- #
 # choose the requested modbus protocol
 # --------------------------------------------------------------------------- #
 
-#from pymodbus.client.async import ModbusUdpClientProtocol
+#from pymodbus.client.asynchronous import ModbusUdpClientProtocol
 
 # --------------------------------------------------------------------------- #
 # configure the client logging
