+++ examples/functional/asynchronous_tcp_client.py
@@ -2,7 +2,7 @@
 import unittest
 from twisted.internet import reactor, protocol
 from pymodbus.constants import Defaults
-from pymodbus.client.async import ModbusClientProtocol
+from pymodbus.client.asynchronous import ModbusClientProtocol
 from base_runner import Runner
 
 class AsynchronousTcpClient(Runner, unittest.TestCase):
