+++ pymodbus/client/asynchronous/__init__.py
@@ -4,21 +4,21 @@ Async Modbus Client implementation based on Twisted, tornado and asyncio
 
 Example run::
 
-    from pymodbus.client.async import schedulers
+    from pymodbus.client.asynchronous import schedulers
 
     # Import The clients
 
-    from pymodbus.client.async.tcp import AsyncModbusTCPClient as Client
-    from pymodbus.client.async.serial import AsyncModbusSerialClient as Client
-    from pymodbus.client.async.udp import AsyncModbusUDPClient as Client
+    from pymodbus.client.asynchronous.tcp import AsyncModbusTCPClient as Client
+    from pymodbus.client.asynchronous.serial import AsyncModbusSerialClient as Client
+    from pymodbus.client.asynchronous.udp import AsyncModbusUDPClient as Client
 
-    # For tornado based async client use
+    # For tornado based asynchronous client use
     event_loop, future = Client(schedulers.IO_LOOP, port=5020)
 
-    # For twisted based async client use
+    # For twisted based asynchronous client use
     event_loop, future = Client(schedulers.REACTOR, port=5020)
 
-    # For asyncio based async client use
+    # For asyncio based asynchronous client use
     event_loop, client = Client(schedulers.ASYNC_IO, port=5020)
 
     # Here event_loop is a thread which would control the backend and future is
@@ -73,7 +73,7 @@ class AsyncModbusClientMixin(BaseAsyncModbusClient):
     def __init__(self, host="127.0.0.1", port=Defaults.Port, framer=None,
                  source_address=None, timeout=None, **kwargs):
         """
-        Initializes a Modbus TCP/UDP async client
+        Initializes a Modbus TCP/UDP asynchronous client
         :param host: Host IP address
         :param port: Port
         :param framer: Framer to use
