+++ pymodbus/client/asynchronous/tcp.py
@@ -2,10 +2,10 @@ from __future__ import unicode_literals
 from __future__ import absolute_import
 
 import logging
-from pymodbus.client.async.factory.tcp import get_factory
+from pymodbus.client.asynchronous.factory.tcp import get_factory
 from pymodbus.constants import Defaults
 from pymodbus.compat import IS_PYTHON3, PYTHON_VERSION
-from pymodbus.client.async.schedulers import ASYNC_IO
+from pymodbus.client.asynchronous.schedulers import ASYNC_IO
 
 logger = logging.getLogger(__name__)
 
@@ -16,7 +16,7 @@ class AsyncModbusTCPClient(object):
 
     To use do::
 
-        from pymodbus.client.async.tcp import AsyncModbusTCPClient
+        from pymodbus.client.asynchronous.tcp import AsyncModbusTCPClient
     """
     def __new__(cls, scheduler, host="127.0.0.1", port=Defaults.Port,
                 framer=None, source_address=None, timeout=None, **kwargs):
