+++ pymodbus/client/asynchronous/tornado/__init__.py
@@ -15,8 +15,8 @@ from tornado.ioloop import IOLoop
 from tornado.iostream import IOStream
 from tornado.iostream import BaseIOStream
 
-from pymodbus.client.async import (AsyncModbusClientMixin,
-                                   AsyncModbusSerialClientMixin)
+from pymodbus.client.asynchronous import (AsyncModbusClientMixin,
+                                          AsyncModbusSerialClientMixin)
 from pymodbus.exceptions import ConnectionException
 from pymodbus.compat import byte2int
